% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/null_model.R
\name{null_model}
\alias{null_model}
\title{null_model}
\usage{
null_model(
  fitme,
  data,
  IDs = NULL,
  range = c(-20, 20),
  length.out = 50000,
  seed = 42
)
}
\arguments{
\item{fitme}{(coxme) frailty survival model}

\item{data}{(data.frame) clinical characteristics, with a mandatory
column \code{subject} reporting subjects' IDs.}

\item{IDs}{(chr, default = NULL) character vector of unique IDs
reported in the column \code{subject} of \code{data}. If NULL (default) an
error will be thrown.}

\item{range}{(num, default = c(-20, 20)) reference symmetric interval
for martingale in the form \code{c(-n, n)}.}

\item{length.out}{(int, default = 50000) number of iterations for the
martingale.}

\item{seed}{(int, default = 42) seed for the simulation}
}
\value{
(null_model) the null model containing the formatted
parameters coming from the \code{{coxme}} model.
}
\description{
null_model
}
\examples{
library(coxme)

db <- eortc
db[["subject"]] <- seq_len(nrow(db))

fitme <- coxme(
  Surv(y, uncens) ~ trt + (1|center),
  db
)
nm <- null_model(
  fitme = fitme,
  data = db,
  IDs = db[["subject"]],
  range = c(-10, 10),
  length.out = 5000
)
str(nm, 1)
}
